ml-pipelines/requirements.txt
pandas
numpy
scikit-learn
python-dotenv
duckdb
sqlalchemy
psycopg[binary]
